requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
qdrant-client
sentence-transformers
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
email-validator